pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "janus-visualizations"
version = "1.0.0"
description = "Visualization suite for Project JANUS neuromorphic trading intelligence"
readme = "project_janus/README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Jordan Smith", email = "jordan@example.com"}
]
keywords = ["visualization", "neuromorphic", "trading", "machine-learning", "neuroscience"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    # Core numerical computing
    "numpy",
    "scipy",

    # Visualization
    "matplotlib",

    # Machine learning
    "scikit-learn",
]

[project.optional-dependencies]
# Advanced visualizations
advanced = [
    "umap-learn",
    "torch",
    "seaborn",
]

# Graph visualization
graph = [
    "networkx",
]

# Interactive plotting
interactive = [
    "plotly",
    "bokeh",
]

# Color accessibility
accessibility = [
    "colorspacious",
]

# Data handling
data = [
    "pandas",
    "tqdm",
]

# Image processing
image = [
    "Pillow",
]

# Jupyter support
jupyter = [
    "jupyter",
    "ipywidgets",
]

# All optional dependencies
all = [
    "umap-learn",
    "torch",
    "seaborn",
    "networkx",
    "plotly",
    "bokeh",
    "colorspacious",
    "pandas",
    "tqdm",
    "Pillow",
    "jupyter",
    "ipywidgets",
]

# Development dependencies
dev = [
    # Testing
    "pytest",
    "pytest-cov",
    "pytest-xdist",

    # Linting and formatting
    "ruff",
    "black",
    "isort",

    # Type checking
    "mypy",

    # Pre-commit
    "pre-commit",
]

[project.urls]
Homepage = "https://github.com/nuniesmith/technical_papers"
Repository = "https://github.com/nuniesmith/technical_papers"
Documentation = "https://github.com/nuniesmith/technical_papers/tree/main/project_janus"

# =============================================================================
# Tool Configurations
# =============================================================================

[tool.setuptools.packages.find]
where = ["."]
include = ["project_janus*"]

[tool.setuptools.package-data]
"project_janus.examples" = ["*.py", "*.sh", "*.txt", "*.md"]

# -----------------------------------------------------------------------------
# Ruff - Modern Python linter (replaces flake8, pylint, etc.)
# -----------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py39"
extend-exclude = [
    "*.egg-info",
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "NPY",    # numpy-specific rules
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict parameter
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"test_*.py" = ["E501"]    # allow long lines in tests
"visual_*.py" = ["N806"]  # allow non-lowercase variable names (for matrices)

[tool.ruff.lint.isort]
known-first-party = ["project_janus"]

# -----------------------------------------------------------------------------
# Black - Code formatter
# -----------------------------------------------------------------------------
[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.ruff_cache
  | __pycache__
  | build
  | dist
)/
'''

# -----------------------------------------------------------------------------
# isort - Import sorter
# -----------------------------------------------------------------------------
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true
known_first_party = ["project_janus"]

# -----------------------------------------------------------------------------
# mypy - Type checker
# -----------------------------------------------------------------------------
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Too strict for scientific computing
ignore_missing_imports = true  # Many scientific packages lack types
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "numpy.*",
    "scipy.*",
    "matplotlib.*",
    "sklearn.*",
    "umap.*",
    "torch.*",
    "seaborn.*",
    "networkx.*",
    "plotly.*",
    "PIL.*",
]
ignore_missing_imports = true

# -----------------------------------------------------------------------------
# pytest - Testing framework
# -----------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=project_janus",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "visual: marks tests that generate visual output",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# -----------------------------------------------------------------------------
# Coverage - Code coverage
# -----------------------------------------------------------------------------
[tool.coverage.run]
source = ["project_janus"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
